requirements.txt
Django==3.0.8
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
pylint==2.6.0
autopep8==1.5.4
Pillow==8.1.0
mysqlclient==2.0.1
# Needed for script to put excel into db
pandas==1.2.2
mysql-connector-python==8.0.23
xlrd==2.0.1
openpyxl==3.0.6
django-cors-headers==3.7.0
# End of script requirements

